requirements.txt
psycopg2==2.5.2
Django==1.6.5
beautifulsoup4==4.3.2
mock==1.0.1
pytz==2014.4
requests==2.3.0
django-endless-pagination==2.0
django-filter==0.7
